+++ data-api/open-api-specification.yml
@@ -225,11 +225,6 @@ paths:
       responses:
         '200':
           description: 'Success'
-          content:
-            application/octet-stream:
-              schema:
-                type: string
-                format: byte
         '400':
           description: 'Bad request'
         '403':
